+++ cadwyn/exceptions.py
@@ -1,17 +1,4 @@
 from fastapi.routing import APIRoute
-from verselect.exceptions import AppCreationError
-
-__all__ = [
-    "AppCreationError",
-    "CadwynError",
-    "LintingError",
-    "CodeGenerationError",
-    "InvalidGenerationInstructionError",
-    "RouterGenerationError",
-    "RouteAlreadyExistsError",
-    "CadwynStructureError",
-    "ModuleIsNotVersionedError",
-]
 
 
 class CadwynError(Exception):